jpms: Refactor handlebars.caffeine
